requirements.txt
streamlit
pandas
snowflake-connector-python
configparser
matplotlib
numpy
plotly
sqlalchemy